+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/contract/ContractManagementController.java
@@ -1,6 +1,6 @@
 package fourservings_fiveservings.insurance_system_be.team.contract;
 
-import fourservings_fiveservings.insurance_system_be.common.dto.RequestVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.RequestVO;
 import fourservings_fiveservings.insurance_system_be.team.TeamController;
 
 public class ContractManagementController implements TeamController {
