+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/contract/ContractManagementTeamMethod.java
@@ -2,8 +2,8 @@ package fourservings_fiveservings.insurance_system_be.team.contract;
 
 import static fourservings_fiveservings.insurance_system_be.user.customer.CustomerList.CUSTOMER_MANAGER;
 
-import fourservings_fiveservings.insurance_system_be.common.dto.RequestVO;
-import fourservings_fiveservings.insurance_system_be.common.dto.ResponseVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.RequestVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.ResponseVO;
 import fourservings_fiveservings.insurance_system_be.insurance.apply.InsuranceApplication;
 import fourservings_fiveservings.insurance_system_be.insurance.apply.InsuranceApplyList;
 import fourservings_fiveservings.insurance_system_be.team.TeamMethod;
