+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/contract/ContractManagementView.java
@@ -1,7 +1,7 @@
 package fourservings_fiveservings.insurance_system_be.team.contract;
 
 
-import fourservings_fiveservings.insurance_system_be.common.dto.RequestVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.RequestVO;
 import fourservings_fiveservings.insurance_system_be.common.view.DepartmentView;
 
 public class ContractManagementView extends DepartmentView {
