+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/plan/inspection/InsuranceInspectionTeamMethod.java
@@ -1,7 +1,7 @@
 package fourservings_fiveservings.insurance_system_be.team.plan.inspection;
 
-import fourservings_fiveservings.insurance_system_be.common.dto.RequestVO;
-import fourservings_fiveservings.insurance_system_be.common.dto.ResponseVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.RequestVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.ResponseVO;
 import fourservings_fiveservings.insurance_system_be.team.TeamMethod;
 
 public class InsuranceInspectionTeamMethod extends TeamMethod {
