+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/plan/inspection/InsuranceInspectionView.java
@@ -10,8 +10,8 @@ import static fourservings_fiveservings.insurance_system_be.insurance.reward.Rew
 import static fourservings_fiveservings.insurance_system_be.insurance.reward.RewardConstant.MONTH_PAYMENT_FEE;
 import static fourservings_fiveservings.insurance_system_be.user.customer.CustomerView.INSURANCE_NUMBER;
 
-import fourservings_fiveservings.insurance_system_be.common.dto.RequestVO;
-import fourservings_fiveservings.insurance_system_be.common.dto.ResponseVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.RequestVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.ResponseVO;
 import fourservings_fiveservings.insurance_system_be.common.view.DepartmentView;
 import fourservings_fiveservings.insurance_system_be.team.plan.design.model.proposal.ProposalConstant;
