+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/salary/SalaryManagementTeamMethod.java
@@ -1,7 +1,7 @@
 package fourservings_fiveservings.insurance_system_be.team.salary;
 
-import fourservings_fiveservings.insurance_system_be.common.dto.RequestVO;
-import fourservings_fiveservings.insurance_system_be.common.dto.ResponseVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.RequestVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.ResponseVO;
 import fourservings_fiveservings.insurance_system_be.team.TeamMethod;
 
 public class SalaryManagementTeamMethod extends TeamMethod {
