+++ src/main/java/fourservings_fiveservings/insurance_system_be/user/User.java
@@ -13,11 +13,9 @@ import jakarta.persistence.InheritanceType;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
 
 @Getter
 @Entity
-
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Inheritance(strategy = InheritanceType.JOINED)
 @DiscriminatorColumn(name = "ENTITY_TYPE")
@@ -44,7 +42,8 @@ public abstract class User extends BaseEntity {
 
     public UserType userType;
 
-    protected User(String loginId, String password, String email, String phoneNumber, Address address,
+    protected User(String loginId, String password, String email, String phoneNumber,
+        Address address,
         String name,
         String birthDay, UserType userType) {
         this.loginId = loginId;
