+++ src/main/java/fourservings_fiveservings/insurance_system_be/user/auth/AuthService.java
@@ -13,8 +13,8 @@ public class AuthService {
     private final UserRepository<User> userRepository;
 
     @Transactional
-    public void signUp(SignUpRequest signUpRequest) {
+    public void signUp(SignUpRequestDto signUpRequestDto) {
         String encodedPassword = "12asdfdsiwyef983e1";
-        userRepository.save(signUpRequest.toEntity(encodedPassword));
+        userRepository.save(signUpRequestDto.toEntity(encodedPassword));
     }
 }
