+++ src/main/java/fourservings_fiveservings/insurance_system_be/user/customer/CustomerView.java
@@ -1,8 +1,8 @@
 package fourservings_fiveservings.insurance_system_be.user.customer;
 
 import fourservings_fiveservings.insurance_system_be.common.AuthUtil;
-import fourservings_fiveservings.insurance_system_be.common.dto.RequestVO;
-import fourservings_fiveservings.insurance_system_be.common.dto.ResponseVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.RequestVO;
+import fourservings_fiveservings.insurance_system_be.common.vo.ResponseVO;
 import fourservings_fiveservings.insurance_system_be.insurance.Insurance;
 import fourservings_fiveservings.insurance_system_be.insurance.InsuranceType;
 import fourservings_fiveservings.insurance_system_be.team.contract.Contract;
